+++ domain_alias/src/Entity/DomainAlias.php
@@ -100,4 +100,13 @@ class DomainAlias extends ConfigEntityBase implements DomainAliasInterface {
     return $this->redirect;
   }
 
+  /**
+   * {@inheritdoc}
+   */
+  public function postSave(EntityStorageInterface $storage, $update = TRUE) {
+    parent::postSave($storage, $update);
+    // Invalidate cache tags relevant to domains.
+    \Drupal::service('cache_tags.invalidator')->invalidateTags(['rendered', 'url.site']);
+  }
+
 }